+++ fme-android/src/main/java/com/vwo/packages/segmentation_evaluator/enums/SegmentOperandRegexEnum.kt
@@ -30,8 +30,9 @@ enum class SegmentOperandRegexEnum(val regex: String) {
     REGEX_MATCH("^regex\\((.*)\\)"),
     STARTING_STAR("^\\*"),
     ENDING_STAR("\\*$"),
-    GREATER_THAN_MATCH("^gt\\((\\d+\\.?\\d*|\\.\\d+)\\)"),
-    GREATER_THAN_EQUAL_TO_MATCH("^gte\\((\\d+\\.?\\d*|\\.\\d+)\\)"),
-    LESS_THAN_MATCH("^lt\\((\\d+\\.?\\d*|\\.\\d+)\\)"),
-    LESS_THAN_EQUAL_TO_MATCH("^lte\\((\\d+\\.?\\d*|\\.\\d+)\\)");
+    //For support multiple decimal points for semantic version comparision
+    GREATER_THAN_MATCH("^gt\\(([\\d.]+)\\)"),
+    GREATER_THAN_EQUAL_TO_MATCH("^gte\\(([\\d.]+)\\)"),
+    LESS_THAN_MATCH("^lt\\(([\\d.]+)\\)"),
+    LESS_THAN_EQUAL_TO_MATCH("^lte\\(([\\d.]+)\\)");
 }