+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[1.7.0] - 2025-08-05
+[1.7.0] - 2025-08-10
 
 ### Added
 
-- Added support for sending a one-time SDK initialization event to VWO server as part of health-check milestones.
+-  Attribute Support: Use custom variables in pre/post-segmentation.
 
+[1.7.0] - 2025-08-05
+
+- Added support for sending a one-time SDK initialization event to VWO server as part of health-check milestones.
 
 [1.6.4] - 2025-07-25